govsite/requirements.txt
Django==6.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-simple-captcha==0.6.0
